Test for each type being  passed in to the Helper.GetParameters method
created all test for types being passed in and checked in method. 
method has a  Date check yet no implemented calling method passes in any date, so created a custom test object that contained a datetime.
Method when converting convers all values to lower case, however NewtonJson for some reason changes UtteranceSplit to the name utt_Split it the only property this is done to the rest are lower case version of the name