Refactor of Deepgram Client intialization
refactored  internal DeepgramClient initialization as if ,s were in appsetting,json or setting.json they were not being read. changes  needed to get it working -

removed configuration.cs
created AppSettingHelper.cs to handle interacting with the file system and getting any values stored in json file
credential initialization moved from deepgramClient.cs to ConfigureCredentials.cs.
ConfigureCredentials takes care of finding and cleaning the values from either appsettings or credentials provide via property or constructor.
moved the managing of creating the HTTPclient to HttpClientHelper, requestPerSecond is now trying to find a value in the json files(handled in AppSettingsHelper.cs) were it was not. With the previous implementation you were creating a lot of overhead for a HTTPclient that was being created every time it was called, it has now been cleaned up

test has been added to test all the changes. 

one last thing with the credentials being cleaned during initialization there was no need for the  cleancredentials class it was just creating a lot of overhead with conversion.

Last note - having the SDK look for the appsetting and setting folder in the project solution consuming the SDK is a bit of a security risk - would recommend that Credentials should only be set during creation of the DeepgramClient or thought the Credentials client(sould really be internal only but it currently public)- would result in the removal of 100+ lines of code. 
Also it is unlikely that consumers will be storing sensitive keys in settings or appsettings ,y will most likely be in a usersecret.json or in a key vault. It is a trivial matter for consumers of the SDK to retrieve the credentials from wherever they have stored them and supply them to the constructor