feat: refactor out json files access (#78)
* HelperTests

* resturcture and .net 7.0

brought structure of solution inline with other language Sdks

* Json file reading removed

Remove the code that was trying to read from the Settings.json appsettings.json files 
Refactored out the Clean credentials this is all now being done when the credentials are passed to the constructor
Updated readme to reflect the changes

* refactored out Uri creations 

added a Uri extensions class's and replaced LiveTranscription.GetWSSUriWithQuerystring & ApiRequest.GetUriWithQuerystring with calls to it. reducing code duplication. 
Also the UriExtension method takes care of adding "/v1" uri segement to the uri removing the need for it to be part of all the calls to ApiRequest and  GetWSSUriWithQuerystring

* Correct failing logic due to typos

typos in the two files were causing logic to fail