Refactor AbstractRestClient
Refactored the AbstractRestClient so that it can either accept a HttpClient or a IHttpClientFactory. if  a HttpClient is passed in it is stored in the parameter ExternalHttpClient. During a Rest call the methods will checker wheter a HttpClient or a IHttpClientFactory is present then assign or create HttpClient that will make the rest call, this allows for a timeout to be set at anytime. this will address both issues -
Timeout -https://github.com/deepgram/deepgram-dotnet-sdk/issues/127
Socket Exhaustion - https://github.com/deepgram/deepgram-dotnet-sdk/issues/148