UpdateMethod calls
rename the client method to have Async at the end of there name.

all rest client now have 2  constructors, the default which does not take a DeepgramClientOptions and one that does.

setting the proxy for HttpClient need to be done at time of creation. so doing this internally would mean managing the lifecycle of the HttpClient and would bloat the size of the sdk. It is much more practical to get the client to set the proxy when adding the HttpClient to services(readme update to show how this is done). Since it is most likely that when a proxy is needed it will be made for a named client. so internally when the create client method is called it needs to know the name of the  named client, so this this has been added to the DeepgramClientOptions.