RemovedDependency injetion
In trying to make the SDK work with or without Dependency injection as creating a over complicated implementation. If a User wants to  use Dependency injection the can easily wrap the client the need and injected that themselves.
Refactored HttpClient Creation to use a internal service collection and IHttpClientFactory
Timeout needs to able to be adjusted inbetween method calls as if a user is not using a  callback method on the prerecoreded client the HttpTimeout was being tripped quickly this becomes apparent when you run longer input files and increase the number of Schema options