Formatting of Extra Property
to address git hub issues
https://github.com/deepgram/deepgram-dotnet-sdk/issues/200
https://github.com/deepgram/deepgram-dotnet-sdk/issues/201

implementation  to properly format the Extra Property of the LiveSchema and PrerecordedSchema for the QueryString of request.
The implementation in the QueryParameterUtil class, in the URl Encode methods switch statement. the format does Not change the casing of the key's or value's as the casing may have significance to the user